+++ src/main/java/com/cognitect/transit/impl/MsgpackEmitter.java
@@ -4,12 +4,12 @@
 package com.cognitect.transit.impl;
 
 import com.cognitect.transit.WriteHandler;
-import org.apache.commons.codec.binary.Base64;
 import org.msgpack.packer.Packer;
 
 import java.io.IOException;
 import java.util.Map;
 import java.util.function.Function;
+import java.util.Base64;
 
 public class MsgpackEmitter extends AbstractEmitter {
 
@@ -96,7 +96,7 @@ public class MsgpackEmitter extends AbstractEmitter {
 
     @Override
     public void emitBinary(Object b, boolean asMapKey, WriteCache cache) throws Exception {
-        byte[] encodedBytes = Base64.encodeBase64((byte[])b);
+        byte[] encodedBytes = Base64.getEncoder().encode((byte[])b);
         emitString(Constants.ESC_STR, "b", new String(encodedBytes), asMapKey, cache);
     }